@rickstaa 2024-05-05
Followed up by https://github.com/livepeer/ai-worker/pull/75.